LayoutTests/fast/dom/NodeList/nodelist-item-call-as-function-expected.txt

This tests that items in a NodeList cannot be called indexed using [[Call]].

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList(0) threw exception TypeError: '[object NodeList]' is not a function (evaluating 'nodeList(0)').
PASS successfullyParsed is true

TEST COMPLETE
